Login UI
This commit adds the ui classes that allow the user to enter a username and password. This is basic functionality to accept input of certain type to access a new page.
There is also Status Message methods that can display pop up windows that display a message or return an input of yes or no.